UI Improvements

+++ examples/event_chain/application/models.py
@@ -261,7 +261,7 @@ class EventAssetState:
             self.event_info.start_time,
             self.event_info.end_time,
         ).days
-        self.display_price = self.event_info.ticket_price / 1e+16
+        self.display_price = int(self.event_info.ticket_price / 1e+16)
 
     def get_next_ticket(self):
         if not self.tickets:
